fix :fire: backend: refatoracao mapeamento backend/banco
