--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/DespesaMensalDAO.java
@@ -1,9 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import Tcoins.tcoins.entities.DespesaMensal;
-
-public interface DespesaMensalDAO extends JpaRepository<DespesaMensal, Long> {
-
-}
