--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/LojaDAO.java
@@ -1,10 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import Tcoins.tcoins.entities.Loja;
-
-public interface LojaDAO extends JpaRepository<Loja, Long> {
-	
-	public Loja findByLoja(Loja loja);
-}
