--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/LojaRamosDAO.java
@@ -1,8 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import Tcoins.tcoins.entities.LojaRamos;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LojaRamosDAO extends JpaRepository<LojaRamos, Long> {
-
-}
