--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/PedidoDAO.java
@@ -1,9 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import Tcoins.tcoins.entities.Pedido;
-
-public interface PedidoDAO extends JpaRepository<Pedido, Long> {
-
-}
