--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/PedidoProdutoDAO.java
@@ -1,9 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import Tcoins.tcoins.entities.PedidoProduto;
-
-public interface PedidoProdutoDAO extends JpaRepository<PedidoProduto, Long> {
-
-}
