--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/PlanosDAO.java
@@ -1,7 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import Tcoins.tcoins.entities.Planos;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PlanosDAO extends JpaRepository<Planos, Long> {
-}
