--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/ProdutoDAO.java
@@ -1,7 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import Tcoins.tcoins.entities.Produto;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProdutoDAO extends JpaRepository<Produto,Long> {
-}
