--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/ReceitaMensalDAO.java
@@ -1,8 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import Tcoins.tcoins.entities.ReceitaMensal;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReceitaMensalDAO extends JpaRepository<ReceitaMensal,Long> {
-
-}
