--- tcoins-backend/src/main/java/Tcoins/tcoins/DAO/UserPlanoDAO.java
@@ -1,7 +1,0 @@
-package Tcoins.tcoins.DAO;
-
-import Tcoins.tcoins.entities.UserPlano;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserPlanoDAO extends JpaRepository<UserPlano,Long> {
-}
