+++ tcoins-backend/src/main/java/br/ifpe/tcoins/TcoinsApplication.java
@@ -1,4 +1,4 @@
-package Tcoins.tcoins;
+package br.ifpe.tcoins;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
