+++ tcoins-backend/src/main/java/br/ifpe/tcoins/repository/LojaRamosRepository.java
@@ -1,0 +1,9 @@
+package br.ifpe.tcoins.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import br.ifpe.tcoins.model.LojaRamos;
+
+public interface LojaRamosRepository extends JpaRepository<LojaRamos, Long> {
+
+}
