+++ tcoins-backend/src/main/java/br/ifpe/tcoins/repository/LojaRepository.java
@@ -1,0 +1,10 @@
+package br.ifpe.tcoins.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import br.ifpe.tcoins.model.Loja;
+
+public interface LojaRepository extends JpaRepository<Loja, Long> {
+	
+//	public Loja findByLoja(Loja loja);
+}
