+++ tcoins-backend/src/main/java/br/ifpe/tcoins/repository/PedidoProdutoRepository.java
@@ -1,0 +1,9 @@
+package br.ifpe.tcoins.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import br.ifpe.tcoins.model.PedidoProduto;
+
+public interface PedidoProdutoRepository extends JpaRepository<PedidoProduto, Long> {
+
+}
