+++ tcoins-backend/src/main/java/br/ifpe/tcoins/repository/ProdutoRepository.java
@@ -1,0 +1,9 @@
+package br.ifpe.tcoins.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import br.ifpe.tcoins.model.Produto;
+
+public interface ProdutoRepository extends JpaRepository<Produto,Long> {
+	
+}
