+++ tcoins-backend/src/main/resources/application.properties
@@ -1,7 +1,6 @@
-spring.datasource.url=jjdbc:mariadb://127.0.0.1:3306/tcoins?autoReconnect=true&useSSL=false&characterEncoding=cp1252
-spring.datasource.username=joao
-spring.datasource.password=BHU*nji9
-spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
+spring.datasource.url=jdbc:mysql://127.0.0.1:3306/tcoin2?autoReconnect=true&useSSL=false&characterEncoding=cp1252
+spring.datasource.username=root
+spring.datasource.password=root
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=false
 spring.jpa.generate-ddl=true
